Change type of radius

+++ plasmapy/analysis/time_series/running_moments.py
@@ -17,7 +17,7 @@ from collections import namedtuple
 _run_moment_tuple = namedtuple("Running_Moment", ["run_moment", "time"])
 
 
-def running_mean(signal, radius: float):
+def running_mean(signal, radius: int):
     """
     Calculate the running mean of a sequence.
 
@@ -62,7 +62,7 @@ def running_mean(signal, radius: float):
     return run_mean[window - 1 :] / window
 
 
-def running_moment(signal, radius: float, moment=1, time=None):
+def running_moment(signal, radius: int, moment=1, time=None):
     """
     Calculate either the running mean, standard deviation, skewness or
     excess kurtosis of a sequence.
